otorisasi

+++ app/category/controller.js
@@ -1,4 +1,5 @@
 const Category = require("./model");
+const { policyFor } = require("../policy");
 
 async function index(req, res, next) {
   try {
@@ -11,6 +12,14 @@ async function index(req, res, next) {
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("create", "Category")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memilik akses untuk membuat kategori",
+      });
+    }
+
     const payload = req.body;
     const category = new Category(payload);
     await category.save();
@@ -29,6 +38,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("update", "Category")) {
+      // <-- can update Category
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate kategori`,
+      });
+    }
     const payload = req.body;
     const category = await Category.findOneAndUpdate(
       { _id: req.params.id },
@@ -50,6 +68,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("delete", "Category")) {
+      // <-- can delete Category
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus kategori`,
+      });
+    }
     const category = await Category.findOneAndDelete({ _id: req.params.id });
     return res.json(category);
   } catch (error) {
