+++ app/product/controller.js
@@ -5,6 +5,7 @@ const config = require("../config");
 const fs = require("fs");
 const path = require("path");
 const { error } = require("console");
+const { policyFor } = require("../policy");
 
 async function index(req, res, next) {
   try {
@@ -52,6 +53,15 @@ async function index(req, res, next) {
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("create", "Product")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk membuat produk",
+      });
+    }
+
     let payload = req.body;
 
     if (payload.tags && payload.tags.length) {
@@ -125,6 +135,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    const policy = policyFor(req.user);
+
+    if (!policy.can("update", "Product")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk mengupdate produk",
+      });
+    }
+
     let payload = req.body;
 
     if (payload.tags && payload.tags.length) {
@@ -209,6 +228,13 @@ async function update(req, res, next) {
 
 async function destory(req, res, next) {
   try {
+    const policy = policyFor(req.user);
+    if (!policy.can("delete", "Product")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk menghapus produk",
+      });
+    }
     const product = await Product.findOneAndDelete({ _id: req.params.id });
     const currentImg = `${config.rootPath}/public/upload/${product.image_url}`;
     if (fs.existsSync) {
