+++ app/tag/controller.js
@@ -1,4 +1,5 @@
 const Tag = require("./model");
+const { policyFor } = require("../policy");
 
 async function index(req, res, next) {
   try {
@@ -11,6 +12,15 @@ async function index(req, res, next) {
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("create", "Tag")) {
+      // <-- can create Tag
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat tag`,
+      });
+    }
     const payload = req.body;
     const tag = new Tag(payload);
     await tag.save();
@@ -29,6 +39,16 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("update", "Tag")) {
+      // <-- can update Tag
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate tag`,
+      });
+    }
+
     const payload = req.body;
     const tag = await Tag.findOneAndUpdate({ _id: req.params.id }, payload, {
       new: true,
@@ -49,6 +69,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("delete", "Tag")) {
+      // <-- can delete Tag
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus tag`,
+      });
+    }
     const tag = await Tag.findOneAndDelete({ _id: req.params.id });
     return res.json(tag);
   } catch (error) {
